requirements.txt
streamlit
PyPDF2
pytesseract
pdf2image
elasticsearch
openai
pillow
PyYAML
